+++ replay/data_preparator.py
@@ -1,8 +1,10 @@
 """
-Contains classes ``DataPreparator`` and ``CatFeaturesTransformer``.
+Contains classes for data preparation and categorical features transformation.
 ``DataPreparator`` is used to transform DataFrames to a library format.
-
-`CatFeaturesTransformer`` transforms categorical features with one-hot encoding.
+``Indexed`` is used to convert user and item ids to numeric format.
+``CatFeaturesTransformer`` transforms categorical features with one-hot encoding.
+``ToNumericFeatureTransformer`` leaves only numerical features
+by one-hot encoding of some features and deleting the others.
 """
 import string
 from typing import Dict, List, Optional
@@ -10,9 +12,10 @@ from typing import Dict, List, Optional
 from pyspark.ml.feature import StringIndexerModel, IndexToString, StringIndexer
 from pyspark.sql import DataFrame
 from pyspark.sql import functions as sf
-
+from pyspark.sql.types import NumericType
 
 from replay.constants import AnyDataFrame
+from replay.session_handler import State
 from replay.utils import convert2spark
 
 
@@ -148,65 +151,6 @@ class Indexer:  # pylint: disable=too-many-instance-attributes
             inv_indexer.setLabels(new_labels)
 
 
-class CatFeaturesTransformer:
-    """Transform categorical features in ``cat_cols_list``
-    with one-hot encoding and remove original columns."""
-
-    def __init__(
-        self,
-        cat_cols_list: List,
-        alias: str = "ohe",
-    ):
-        """
-        :param cat_cols_list: list of categorical columns
-        :param alias: prefix for one-hot encoding columns
-        """
-        self.cat_cols_list = cat_cols_list
-        self.expressions_list = []
-        self.alias = alias
-
-    def fit(self, spark_df: Optional[DataFrame]) -> None:
-        """
-        Save categories for each column
-        :param spark_df: Spark DataFrame with features
-        """
-        if spark_df is None:
-            return
-
-        cat_feat_values_dict = {
-            name: (
-                spark_df.select(sf.collect_set(sf.col(name))).collect()[0][0]
-            )
-            for name in self.cat_cols_list
-        }
-        self.expressions_list = [
-            sf.when(sf.col(col_name) == cur_name, 1)
-            .otherwise(0)
-            .alias(
-                f"""{self.alias}_{col_name}_{str(cur_name).translate(
-                        str.maketrans(
-                            "", "", string.punctuation + string.whitespace
-                        )
-                    )[:30]}"""
-            )
-            for col_name, col_values in cat_feat_values_dict.items()
-            for cur_name in col_values
-        ]
-
-    def transform(self, spark_df: Optional[DataFrame]):
-        """
-        Transform categorical columns.
-        If there are any new categories that were not present at fit stage, they will be ignored.
-        :param spark_df: feature DataFrame
-        :return: transformed DataFrame
-        """
-        if spark_df is None:
-            return None
-        return spark_df.select(*spark_df.columns, *self.expressions_list).drop(
-            *self.cat_cols_list
-        )
-
-
 class DataPreparator:
     """
     Convert pandas DataFrame to Spark, rename columns and apply indexer.
@@ -277,3 +221,185 @@ class DataPreparator:
         :return: DataFrame with ids
         """
         return self.indexer.inverse_transform(df)
+
+
+class CatFeaturesTransformer:
+    """Transform categorical features in ``cat_cols_list``
+    with one-hot encoding and remove original columns."""
+
+    def __init__(
+        self,
+        cat_cols_list: List,
+        alias: str = "ohe",
+    ):
+        """
+        :param cat_cols_list: list of categorical columns
+        :param alias: prefix for one-hot encoding columns
+        """
+        self.cat_cols_list = cat_cols_list
+        self.expressions_list = []
+        self.alias = alias
+
+    def fit(self, spark_df: Optional[DataFrame]) -> None:
+        """
+        Save categories for each column
+        :param spark_df: Spark DataFrame with features
+        """
+        if spark_df is None:
+            return
+
+        cat_feat_values_dict = {
+            name: (
+                spark_df.select(sf.collect_set(sf.col(name))).collect()[0][0]
+            )
+            for name in self.cat_cols_list
+        }
+        self.expressions_list = [
+            sf.when(sf.col(col_name) == cur_name, 1)
+            .otherwise(0)
+            .alias(
+                f"""{self.alias}_{col_name}_{str(cur_name).translate(
+                        str.maketrans(
+                            "", "", string.punctuation + string.whitespace
+                        )
+                    )[:30]}"""
+            )
+            for col_name, col_values in cat_feat_values_dict.items()
+            for cur_name in col_values
+        ]
+
+    def transform(self, spark_df: Optional[DataFrame]):
+        """
+        Transform categorical columns.
+        If there are any new categories that were not present at fit stage, they will be ignored.
+        :param spark_df: feature DataFrame
+        :return: transformed DataFrame
+        """
+        if spark_df is None:
+            return None
+        return spark_df.select(*spark_df.columns, *self.expressions_list).drop(
+            *self.cat_cols_list
+        )
+
+
+class ToNumericFeatureTransformer:
+    """Transform user/item features to numeric types:
+    - numeric features stays as is
+    - categorical features:
+        if threshold is defined:
+            - all non-numeric columns with less unique values than threshold are one-hot encoded
+            - remaining columns are dropped
+        else all non-numeric columns are one-hot encoded
+    """
+
+    cat_feat_transformer: Optional[CatFeaturesTransformer]
+    cols_to_ohe: Optional[List]
+    cols_to_del: Optional[List]
+    all_columns: Optional[List]
+
+    def __init__(self, threshold: Optional[int] = 100):
+        self.threshold = threshold
+        self.fitted = False
+
+    def fit(self, features: Optional[DataFrame]) -> None:
+        """
+        Determine categorical columns for one-hot encoding.
+        Non categorical columns with more values than threshold will be deleted.
+        Saves categories for each column.
+        :param features: input DataFrame
+        """
+        self.cat_feat_transformer = None
+        self.cols_to_del = []
+        self.fitted = True
+
+        if features is None:
+            self.all_columns = None
+            return
+
+        self.all_columns = sorted(features.columns)
+        idx_cols_set = {"user_idx", "item_idx", "user_id", "item_id"}
+
+        spark_df_non_numeric_cols = [
+            col
+            for col in features.columns
+            if (not isinstance(features.schema[col].dataType, NumericType))
+            and (col not in idx_cols_set)
+        ]
+
+        # numeric only
+        if len(spark_df_non_numeric_cols) == 0:
+            self.cols_to_ohe = []
+            return
+
+        if self.threshold is None:
+            self.cols_to_ohe = spark_df_non_numeric_cols
+        else:
+            counts_pd = (
+                features.agg(
+                    *[
+                        sf.approx_count_distinct(sf.col(c)).alias(c)
+                        for c in spark_df_non_numeric_cols
+                    ]
+                )
+                .toPandas()
+                .T
+            )
+            self.cols_to_ohe = (
+                counts_pd[counts_pd[0] <= self.threshold]
+            ).index.values
+
+            self.cols_to_del = [
+                col
+                for col in spark_df_non_numeric_cols
+                if col not in set(self.cols_to_ohe)
+            ]
+
+            if self.cols_to_del:
+                State().logger.warning(
+                    "%s columns contain more that threshold unique "
+                    "values and will be deleted",
+                    self.cols_to_del,
+                )
+
+        if len(self.cols_to_ohe) > 0:
+            self.cat_feat_transformer = CatFeaturesTransformer(
+                cat_cols_list=self.cols_to_ohe
+            )
+            self.cat_feat_transformer.fit(features.drop(*self.cols_to_del))
+
+    def transform(self, spark_df: Optional[DataFrame]) -> Optional[DataFrame]:
+        """
+        Transform categorical features.
+        Use one hot encoding for columns with the amount of unique values smaller
+        than threshold and delete other columns.
+        :param spark_df: input DataFrame
+        :return: processed DataFrame
+        """
+        if not self.fitted:
+            raise AttributeError("Call fit before running transform")
+
+        if spark_df is None or self.all_columns is None:
+            return None
+
+        if self.cat_feat_transformer is None:
+            return spark_df.drop(*self.cols_to_del)
+
+        if sorted(spark_df.columns) != self.all_columns:
+            raise ValueError(
+                f"Columns from fit do not match "
+                f"columns in transform. "
+                f"Fit columns: {self.all_columns},"
+                f"Transform columns: {spark_df.columns}"
+            )
+
+        return self.cat_feat_transformer.transform(
+            spark_df.drop(*self.cols_to_del)
+        )
+
+    def fit_transform(self, spark_df: DataFrame) -> DataFrame:
+        """
+        :param spark_df: input DataFrame
+        :return: output DataFrame
+        """
+        self.fit(spark_df)
+        return self.transform(spark_df)
