+++ replay/scenarios/__init__.py
@@ -4,6 +4,3 @@ Scenarios are a series of actions for recommendations
 from replay.scenarios.fallback import Fallback
 from replay.scenarios.basescenario import BaseScenario
 from replay.scenarios.two_stages.two_stages_scenario import TwoStagesScenario
-from replay.scenarios.two_stages.feature_processor import (
-    HistoryBasedFeaturesProcessor,
-)
